strands-agents/enterprise-ai/backend/requirements.txt
# Core Flask Framework
Flask>=2.3.3
Flask-RESTful>=0.3.10
Flask-Cors>=4.0.0

# Environment and Configuration
python-dotenv==1.0.0

# AI and Agent Framework
strands-agents
strands-agents-tools
langchain-mcp-adapters

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
postgres-mcp

# Model Context Protocol
mcp>=1.0.0

# Optional: Remove if not using ClickHouse
# clickhouse-connect>=0.8.16
emoji
